Move Enum and CorElementType to shared (#23177)
